Move IpPacketRouterAddress to gateway_client
