+++ nym-vpn-lib/src/mixnet_processor.rs
@@ -14,7 +14,6 @@ use nym_ip_packet_requests::{
 };
 use nym_sdk::mixnet::{InputMessage, MixnetMessageSender, Recipient};
 use nym_task::{connections::TransmissionLane, TaskClient, TaskManager};
-use nym_validator_client::models::DescribedGateway;
 use tokio::task::JoinHandle;
 use tokio::time::timeout;
 use tokio_util::codec::Decoder;
@@ -24,6 +23,7 @@ use tun2::{AbstractDevice, AsyncDevice};
 use crate::{
     connection_monitor::{self, ConnectionStatusEvent},
     error::{Error, Result},
+    gateway_client::IpPacketRouterAddress,
     icmp_connection_beacon,
     mixnet_connect::SharedMixnetClient,
 };
@@ -38,37 +38,6 @@ impl Config {
         Config {
             ip_packet_router_address,
         }
-    }
-}
-
-#[derive(Debug, Copy, Clone)]
-pub struct IpPacketRouterAddress(pub Recipient);
-
-impl IpPacketRouterAddress {
-    pub fn try_from_base58_string(ip_packet_router_nym_address: &str) -> Result<Self> {
-        Ok(Self(
-            Recipient::try_from_base58_string(ip_packet_router_nym_address)
-                .map_err(|_| Error::RecipientFormattingError)?,
-        ))
-    }
-
-    pub fn try_from_described_gateway(gateway: &DescribedGateway) -> Result<Self> {
-        let address = gateway
-            .self_described
-            .clone()
-            .and_then(|described_gateway| described_gateway.ip_packet_router)
-            .map(|ipr| ipr.address)
-            .ok_or(Error::MissingIpPacketRouterAddress)?;
-        Ok(Self(
-            Recipient::try_from_base58_string(address)
-                .map_err(|_| Error::RecipientFormattingError)?,
-        ))
-    }
-}
-
-impl std::fmt::Display for IpPacketRouterAddress {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", self.0)
     }
 }
